split model into tables/views
